refactor: DNSTransport into internal/model
Let's continue moving important interfaces into internal/model.
