+++ internal/model/netx.go
@@ -1,6 +1,7 @@
 package model
 
 import (
+	"context"
 	"net"
 	"syscall"
 )
@@ -58,6 +59,24 @@ type DNSEncoder interface {
 	// On success, this function returns a valid byte array and
 	// a nil error. On failure, we have an error and the byte array is nil.
 	Encode(domain string, qtype uint16, padding bool) ([]byte, error)
+}
+
+// DNSTransport represents an abstract DNS transport.
+type DNSTransport interface {
+	// RoundTrip sends a DNS query and receives the reply.
+	RoundTrip(ctx context.Context, query []byte) (reply []byte, err error)
+
+	// RequiresPadding returns whether this transport needs padding.
+	RequiresPadding() bool
+
+	// Network is the network of the round tripper (e.g. "dot").
+	Network() string
+
+	// Address is the address of the round tripper (e.g. "1.1.1.1:853").
+	Address() string
+
+	// CloseIdleConnections closes idle connections, if any.
+	CloseIdleConnections()
 }
 
 // HTTPSSvc is the reply to an HTTPS DNS query.
