+++ internal/netxlite/serialresolver.go
@@ -26,11 +26,11 @@ type SerialResolver struct {
 	NumTimeouts *atomicx.Int64
 
 	// Txp is the underlying DNS transport.
-	Txp DNSTransport
+	Txp model.DNSTransport
 }
 
 // NewSerialResolver creates a new SerialResolver instance.
-func NewSerialResolver(t DNSTransport) *SerialResolver {
+func NewSerialResolver(t model.DNSTransport) *SerialResolver {
 	return &SerialResolver{
 		Encoder:     &DNSEncoderMiekg{},
 		Decoder:     &DNSDecoderMiekg{},
@@ -40,7 +40,7 @@ func NewSerialResolver(t DNSTransport) *SerialResolver {
 }
 
 // Transport returns the transport being used.
-func (r *SerialResolver) Transport() DNSTransport {
+func (r *SerialResolver) Transport() model.DNSTransport {
 	return r.Txp
 }
